requirements.txt
tk==0.1.0
Pillow==9.2.0
